Cleanup, CSS tweaks
