+++ src/pages/images/MetadataDrawer/ImageMetadataPanel.tsx
@@ -36,7 +36,7 @@ export const ImageMetadataPanel = (props: ImageMetadataPanelProps) => {
   return (
     <PropertyValidation>
       <form 
-        className="flex flex-col justify-between h-full"
+        className="flex flex-col justify-between h-full py-4 px-6"
         onSubmit={onSubmit}>
         <div className="flex flex-col flex-grow">
           <h2 className="leading-relaxed mr-5 mb-8 font-medium">
