--- src/pages/knowledgegraph/Controls/index.ts
@@ -1,1 +1,0 @@
-export * from './Controls';
