--- src/pages/knowledgegraph/DetailsDrawer/index.ts
@@ -1,1 +1,0 @@
-export * from './DetailsDrawer';
