+++ src/pages/knowledgegraph/GraphControls/index.ts
@@ -1,0 +1,1 @@
+export * from './GraphControls';
