+++ src/pages/knowledgegraph/SelectionDetailsDrawer/EntityTypeDetails/EntityTypeDetails.tsx
