+++ src/pages/knowledgegraph/SelectionDetailsDrawer/EntityTypeDetails/index.ts
