+++ src/pages/knowledgegraph/SelectionDetailsDrawer/ImageDetails/AnnotationsTab/AnnotationsTab.tsx
