+++ src/pages/knowledgegraph/SelectionDetailsDrawer/ImageDetails/AnnotationsTab/AnnotationsTabItem.tsx
