+++ src/pages/knowledgegraph/SelectionDetailsDrawer/ImageDetails/AnnotationsTab/index.ts
