+++ src/pages/knowledgegraph/SelectionDetailsDrawer/ImageDetails/ImageDetails.tsx
