+++ src/pages/knowledgegraph/SelectionDetailsDrawer/ImageDetails/MetadataTab/MetadataTab.tsx
