+++ src/pages/knowledgegraph/SelectionDetailsDrawer/ImageDetails/MetadataTab/index.ts
