+++ src/pages/knowledgegraph/SelectionDetailsDrawer/ImageDetails/index.ts
