+++ src/pages/knowledgegraph/SelectionDetailsDrawer/index.ts
@@ -1,0 +1,1 @@
+export * from './SelectionDetailsDrawer';
